AC:移除使用者包含移除舊www目錄

+++ sh/admin_util/.post_deluser.sh
@@ -9,7 +9,15 @@ source /root/admin_util/.env
 sudo deluser --remove-all-files $the_user
 echo "001 使用者目錄移除...成功。"
 
-sudo rm -f $HTML_DIR/.$the_user
-echo "002 使用者WWW連結移除...成功。"
+if [ -d $HTML_DIR/$the_user ]; then
+    sudo rm -f $HTML_DIR/$the_user
+    echo "002.1 使用者WWW連結 $HTML_DIR/$the_user 移除...成功。"
+fi
+
+if [ -d $HTML_DIR/.$the_user ]; then
+    sudo rm -f $HTML_DIR/.$the_user
+    echo "002.2 使用者WWW連結 $HTML_DIR/.$the_user 移除...成功。"
+fi
+
 
 echo "使用者 $the_user 資料移除 完成!"
